fix: comprehensive code quality improvements
- Fixed all TypeScript compilation errors
- Added error handling with try-catch blocks in crypto operations
- Optimized toBase64 performance (loop instead of spread operator)
- Removed unnecessary type assertions
- Applied optional chaining for null checks
- Changed Error to TypeError for type validation
- Fixed trailing whitespace issues
- Enhanced error messages across all modules
- Added crypto availability check
- Fixed Uint8Array/BufferSource type compatibility
- Fixed Date constructor and array sorting issues
- Fixed BlobPart type issues in file operations
- Removed unused sync operation

+++ test/capsule.test.ts
@@ -83,7 +83,7 @@ describe('Capsule Integration', () => {
       metadata: { ...savedMetadata, capsule_id: capsuleId }
     })
 
-    const encryptedPayload = mockPinata.uploadBytes.mock.calls[0][0]
+    const encryptedPayload = mockPinata.uploadBytes.mock.calls[0][0] as Uint8Array
     mockPinata.getBytes.mockResolvedValueOnce(encryptedPayload)
 
     const decrypted = await manager.unlockCapsule({
